refactor(teacher): TeacherClasses sous 300 lignes — éclatement complet (H9)
614 → 243 lignes. Extraction:
- composables/useTeacherClasses.js (chargement + enrichissement + cache + refresh arrière-plan)
- components/teacher/TeacherClassCard.vue (carte présentationnelle, CSS verbatim)
- tests: composable + carte (5 tests verts)
